requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
